store events in encrypted preferences

+++ app/src/main/java/com/bytemedrive/store/StoreRepository.kt
@@ -9,7 +9,7 @@ import io.ktor.client.request.setBody
 
 class StoreRepository(private val httpClient: HttpClient) {
 
-    suspend fun getEncryptedEvents(usernameSha3: String, credentialsSha3: String, offset: Int): List<EncryptedEvent> =
+    suspend fun getEncryptedEvents(usernameSha3: String, credentialsSha3: String, offset: Long): List<EncryptedEvent> =
         httpClient.create().get("customers/${usernameSha3}/events?offset=${offset}") {
             headers { append("Authorization", "Hash $credentialsSha3") }
         }.body()
